perf: Remove inline require call
