+++ server/policies/collection.test.js
@@ -2,7 +2,7 @@
 import { CollectionUser, Collection } from "../models";
 import { buildUser, buildTeam, buildCollection } from "../test/factories";
 import { flushdb } from "../test/support";
-import { serialize } from "./index";
+import serialize from "./serializer";
 
 beforeEach(() => flushdb());
