+++ server/policies/document.test.js
@@ -6,7 +6,7 @@ import {
   buildCollection,
 } from "../test/factories";
 import { flushdb } from "../test/support";
-import { serialize } from "./index";
+import serialize from "./serializer";
 
 beforeEach(() => flushdb());
